SurveyApplication_1.0

+++ src/main/java/pl/surveyapplication/repository/SurveyRepository.java
@@ -6,8 +6,13 @@ import pl.surveyapplication.model.Survey;
 
 import java.util.List;
 
+/**
+ * @author Dawid
+ * @version 1.0
+ * Interface repozytoryjny magazynu wypełnionych ankiet. Pozwala na interakcje pomiędzy aplikacją a bazą danych.
+ * */
 @Repository
-public interface SurveyRepository extends CrudRepository<Survey,Integer> {
+public interface SurveyRepository extends CrudRepository<Survey,Long> {
     @Override
     List<Survey> findAll();
 }
