+++ src/main/java/pl/surveyapplication/repository/UserRepository.java
@@ -6,8 +6,13 @@ import pl.surveyapplication.model.User;
 import java.util.List;
 import java.util.Optional;
 
+/**
+ * @author Dawid
+ * @version 1.0
+ * Interface repozytoryjny użytkowników. Pozwala na interakcje pomiędzy aplikacją a bazą danych.
+ * */
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
     @Override
     List<User> findAll();
